Momo Email Assistant.app/requirements.txt
fastapi
uvicorn
python-dotenv
requests
schedule
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai